% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorkey.R
\name{colorkey}
\alias{colorkey}
\title{colorkey}
\usage{
colorkey(colortext, excludetext)
}
\arguments{
\item{colortext}{Regular expression to match values returned by
\code{colors()}}

\item{excludetext}{Regular expression to match values to be
\emph{excluded} from the list created by filtering on
\code{colortext}.}
}
\value{
By default creates a plot and otherwise returns
    nothing. Assigned output is a character vector of the matching
    color names.
}
\description{
\code{colorkey} produces a plot showing all standard R
    colornames (as returned by \code{colors}) matching a regular
    expression. Matching is case insensitive. If output is assigned
    to a variable, the variable contains the list of colors. By
    default, returns all colors.

\code{colorkey(colortext)}
}
\examples{
\dontrun{

## color names containing "red"
colorkey('blue')
colorkey('blue', excludetext = 'light|dark')
## color names containing "red" but  
## color names starting with "bl" or "gr"
colorkey('^(gr)|^(bl)')

## plots colors starting with "bl" and then either "u" or "a", and
## returns the name in the variable `namelist`
namelist <- colorkey('^bl[ua]')

}
}
