% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{data_summary}
\alias{data_summary}
\alias{num_summaries}
\title{data_summary}
\usage{
data_summary(d, numdigits)
}
\arguments{
\item{numdigits}{number of digits for formatted output}

\item{d}{A data frame}
}
\value{
By default creates a plot and otherwise returns
    nothing. Assigned output is a character vector of the matching
    color names.
}
\description{
\code{data_summary} reports, for a data frame, the
    number of observations, number of missing, class of each
    variable, and summary statistics, including mean, standard
    deviation, and various quantiles, including the minimum,
    median, and maximum.

\code{data_summary(d)}
}
\examples{

data_summary(iris)

}
