% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateformats.R
\name{dateformats}
\alias{dateformats}
\alias{posix_dt_format}
\alias{translate_datestr}
\title{Date formatting strings}
\usage{
posix_dt_format(syms, categories, description_short, return_all)
    translate_datestr(datetimestr)
}
\arguments{
\item{syms}{A vector of POSIX date-time codes, \code{NULL} by
default}

\item{categories}{posix date categories, one or more of `c('date',
'day', 'week', 'month', 'year', 'time', 'date-time',
'formatting').}

\item{description_short}{boolean, whether to return short or long
code descriptions.}

\item{return_all}{boolean, whether to return a data frame of all 36
date-time codes with both short and long descriptions}

\item{datetimestr}{string containing codes to be matched}
}
\value{
Returns a data frame with all or a subset of POSIX
    date-time codes, code categories, and descriptions, all of
    which are extracted from \code{?strftime}.
}
\description{
\code{posix_dt_format} returns all or a portion of a
    data frame describing POSIX date
    strings. \code{translate_datestr} accepts a string and extracts
    the POSIX date codes, returning the portion of the data frame
    documenting those codes.
}
\examples{

posix_dt_format(syms=c('\%B', '\%m', '\%Y', '\%y'))

translate_datestr(datetimestr = '\%Y-\%M-\%d')



}
