% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addl_knitr_engines.R
\name{.soln_notes}
\alias{.soln_notes}
\title{knitr engine for italicizing chunk}
\usage{
.soln_notes(options)
}
\arguments{
\item{options}{This is a list returned from the chunk and contains
all chunk options (these are available as `code$eval`,
`code$echo`, etc.along with the chunk contents (in
options$code). Have to pass chunk options through this list
rather than as arguments to the function.}
}
\description{
Engine to put entire chunk contents (no code) in
    italic. Use for problem set solutions. This engine is
    automatically available when the package is loaded (see
    `zzz.R`). In general the syntax for `knit_engines$set` has to
    match what's in `zzz.R`
}
