% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nobreakspace.R
\name{nbsp}
\alias{nbsp}
\title{Non-breaking spaces}
\usage{
nbsp(str, verb = TRUE)
}
\arguments{
\item{str}{Character string}

\item{verb}{Boolean to quote as verbatim}
}
\value{
Replaces spaces with `&nbsp;` and optionally formats as
    verbatim
}
\description{
Inserts non-breaking space in Rmarkdown. 

\code{nbsp(str, verb=TRUE)}
}
\examples{
\dontrun{

## Create inline hyperlink
s = "File|Import Dataset"
`r nbsp(s)`
`r nbsp(s, verb=FALSE)`
}
}
