% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkglink.R
\name{pkglink}
\alias{pkglink}
\title{pkglink}
\usage{
pkglink(pkgname)
}
\arguments{
\item{pkgname}{Quoted package name}
}
\value{
String containing hyperlink to CRAN page for the given
    package
}
\description{
provides Rmarkdown-formatted CRAN link for a given
    quoted package name

\code{pkglink(pkgname)}
}
\examples{
\dontrun{

## Create inline hyperlink
`r pkglink('dplyr')`
## Create multiple inline hyperlinks, in alphabetical order
`r lapply(sort(c('dplyr', 'ggplot2', 'derivmkts')), pkglink)`
#' ## Create multiple inline hyperlinks in supplied order
`r knitr::combine_words(lapply(c('dplyr', 'ggplot2', 'derivmkts'), pkglink))`
}
}
