% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordle_assist.R
\name{wordle-assist}
\alias{wordle-assist}
\alias{wordle_assist}
\title{wordle-assist}
\usage{
wordle_assist(no = NULL, yes = NULL, words = words5)
}
\arguments{
\item{no}{letters not in the word. These should be specified as a
character string, for example, \code{'aeontfd'}}

\item{yes}{a list of letters included in the word, with positional
information. Correct positions are positive integers and
incorrect positions are negative integers. For example if "a"
is in the second position and not in the fourth, and "y" is in
the fifth, specify \code{yes = list(a = c(2, -4), y = 5)}}

\item{words}{vector of admissible possible words. By default, uses
the length-5 words from the `words` package.}
}
\value{
Returns a vector of 5-letter words meeting supplied match
    criteria
}
\description{
\code{wordle-assist} returns 5-letter words that match
    included and excluded letter patterns.
}
\examples{

wordle_assist(no = 'seuntilmp', yes = list(a = -1, r = -2, o = -3, c = -1, h = -2))


}
