% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/words5.R
\docType{data}
\name{words5}
\alias{words5}
\title{All length-5 words in the `words` package}
\format{
vector of length-5 words
}
\source{
The `words` package:  \cr\cr
\code{words5 <- words::words[words::words$word_length==5, 'word']}
}
\usage{
words5
}
\description{
All length-5 words in the `words` package
}
\keyword{datasets}
