% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/words_subset.R
\name{words_subset}
\alias{words_subset}
\title{words_subset}
\usage{
words_subset(ltrs = NULL, words = words5)
}
\arguments{
\item{ltrs}{letters in the word. These should be specified as a
character string, for example, \code{'aeontfd'}}

\item{words}{vector of admissible possible words. By default, uses
the length-5 words from the `words` package.}
}
\value{
Returns a vector of 5-letter words meeting supplied match
    criteria
}
\description{
\code{words_subset} returns 5-letter words that
    contain specified letters. Error if more than 5 letters are
    supplied
}
\examples{

words_subset(ltrs = 'cawk')


}
