% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorkey.R
\name{colorkey}
\alias{colorkey}
\title{colorkey}
\usage{
colorkey(colortext)
}
\arguments{
\item{colortext}{Regular expression to match values returned by
\code{colors()}}
}
\value{
By default creates a plot and otherwise returns
    nothing. Assigned output is a character vector of the matching
    color names.
}
\description{
\code{colorkey} produces a plot showing all standard R
    colornames (as returned by \code{colors}) matching a regular
    expression. If output is assigned to a variable, the variable
    contains the list of colors.

\code{colorkey(colortext)}
}
\examples{
\dontrun{

## colors containing "red"
colorkey('red')

## color names starting with "bl" or "gr"
colorkey('^(gr)|^(bl)')

## plots colors starting with "bl" and then either "u" or "a", and
## returns the name in the variable `namelist`
namelist <- colorkey('^bl[ua]')

}
}
